fetch most recent aggregate data
a previous fix changed the code to use only the first value instead of a
summation. but the date condition (since) was not removed. this has now
been removed so we only fetch the 3 most recent data points. we also
select the last data point, which is the most recent one.